requirements.txt
beautifulsoup4==4.9.0
lxml==4.2.1
mysql-connector-python==8.0.19
regex==2020.4.4
selenium==3.141.0

